Renamed connector to adapter
