+++ MixerControl-app/adapter/auth_service_adapter.js
