+++ MixerControl-app/routes/components.js
@@ -5,7 +5,7 @@ var helper = require('../services/helper_service');
 var cache = require('../services/cache_middleware');
 
 
-var jms_connector = require('../connectors/juice_machine_service_connector');
+var jms_connector = require('../adapter/juice_machine_service_adapter');
 
 /**
  * Functions
